# Better operator string represenation (take 2)