--- a/akka-stream/src/main/scala/akka/stream/Attributes.scala
+++ b/akka-stream/src/main/scala/akka/stream/Attributes.scala
@@ -21,8 +21,6 @@ import akka.util.{ ByteString, OptionVal }
 import akka.util.JavaDurationConverters._
 import akka.util.LineNumbers
 
-import java.util.concurrent.ConcurrentHashMap
-
 /**
  * Holds attributes which can be used to alter [[akka.stream.scaladsl.Flow]] / [[akka.stream.javadsl.Flow]]
  * or [[akka.stream.scaladsl.GraphDSL]] / [[akka.stream.javadsl.GraphDSL]] materialization.
@@ -309,21 +307,24 @@ object Attributes {
    * Attribute that contains the source location of for example a lambda passed to an operator, useful for example
    * for debugging. Included in the default toString of GraphStageLogic if present
    */
-  final case class SourceLocation(locationName: String) extends Attribute
+  final class SourceLocation(lambda: AnyRef) extends Attribute {
+    lazy val locationName: String = {
+      val locationName = LineNumbers(lambda) match {
+        case LineNumbers.NoSourceInfo           => "unknown"
+        case LineNumbers.UnknownSourceFormat(_) => "unknown"
+        case LineNumbers.SourceFile(filename)   => filename
+        case LineNumbers.SourceFileLines(filename, from, _) =>
+          s"$filename:$from"
+      }
+      s"${lambda.getClass.getPackage.getName}-$locationName"
+    }
+
+    override def toString: String = locationName
+  }
 
   object SourceLocation {
-    private val cache = new ConcurrentHashMap[Class[_], SourceLocation]
-    def forLambda(lambda: AnyRef): SourceLocation =
-      cache.computeIfAbsent(lambda.getClass, { clazz =>
-        def locationName: String = LineNumbers(lambda) match {
-          case LineNumbers.NoSourceInfo           => "unknown"
-          case LineNumbers.UnknownSourceFormat(_) => "unknown"
-          case LineNumbers.SourceFile(filename)   => filename
-          case LineNumbers.SourceFileLines(filename, from, _) =>
-            s"$filename:$from"
-        }
-        SourceLocation(s"${clazz.getPackage.getName}-$locationName")
-      })
+    def forLambda(lambda: AnyRef): SourceLocation = new SourceLocation(lambda)
+
     def stringFrom(attributes: Attributes): String =
       attributes.get[SourceLocation].map(_.locationName).getOrElse("unknown")
   }