--- a/akka-stream/src/main/scala/akka/stream/impl/Stages.scala
+++ b/akka-stream/src/main/scala/akka/stream/impl/Stages.scala
@@ -27,6 +27,7 @@ import akka.stream.Attributes._
     val filterNot = name("filterNot")
     val collect = name("collect")
     val recover = name("recover")
+    val mapError = name("mapError")
     val mapAsync = name("mapAsync")
     val mapAsyncUnordered = name("mapAsyncUnordered")
     val ask = name("ask")