… akka-stream-tests/src/test/scala/akka/stream/impl/GraphStageLogicSpec.scala
… akka-stream-tests/src/test/scala/akka/stream/scaladsl/FromMaterializationSpec.scala
… akka-stream-tests/src/test/scala/akka/stream/scaladsl/SetupSpec.scala