Based on @jrudolph s great idea in #29874 (so supersedes that PR)

Instead of adding user provided lambda source location in the name attribute this PR introduces a new Attribute with source location. The default toString of `GraphStageLogic` has then been updated to include that information if available and relevant debug/error and stream snapshot labels for the operator have been updated to use the toString of `GraphStageLogic`. 

Example of preStart exception from one of the tests for example:
```
Error during preStart in [akka.stream.impl.SetupFlowStage$$anon$2-setup(akka.stream.scaladsl-FromMaterializationSpec.scala:141)]: boom
```

The format is `[logic instance class name]-[name attribute]([source location of lambda])`