@raboof 2021-01-08
an unused import only on Scala 2.13 :thinking: