@johanandren 2021-01-08
I'm a bit worried that this has a too high penalty for materialization, a short run of the MaterializationBenchmark shows from pretty much no slow down to ~15% slower for some cases but I'm not sure about those numbers, need to run the bench some more and double check.