@johanandren 2021-01-11
Re-ran benchmark today. I think the materialization hit is low enough/close enough to noise that we can accept this and deal with any specific performance problems if they arise.


```
With SourceLocation attribute
Benchmark                                                (complexity)   Mode  Cnt       Score       Error  Units   (vs baseline)
MaterializationBenchmark.flow_with_map                              1  thrpt    5  239992,282 ± 28630,943  ops/s    96%   
MaterializationBenchmark.flow_with_map                             10  thrpt    5  142910,329 ±  2849,390  ops/s    93%
MaterializationBenchmark.flow_with_map                            100  thrpt    5   38539,328 ±  1061,710  ops/s    86%
MaterializationBenchmark.graph_with_imported_flow                   1  thrpt    5  241854,339 ± 35103,036  ops/s   106%
MaterializationBenchmark.graph_with_imported_flow                  10  thrpt    5  129178,802 ±  1154,565  ops/s    96%
MaterializationBenchmark.graph_with_imported_flow                 100  thrpt    5   31139,182 ±  1062,769  ops/s    92%
MaterializationBenchmark.graph_with_junctions_gradual               1  thrpt    5  239125,225 ± 39726,150  ops/s   103%
MaterializationBenchmark.graph_with_junctions_gradual              10  thrpt    5  103421,596 ±  2063,545  ops/s    99%
MaterializationBenchmark.graph_with_junctions_gradual             100  thrpt    5    9219,574 ±   279,067  ops/s    97%
MaterializationBenchmark.graph_with_junctions_immediate             1  thrpt    5  208140,651 ± 11779,395  ops/s    98%
MaterializationBenchmark.graph_with_junctions_immediate            10  thrpt    5  156420,103 ±  5372,413  ops/s    96%
MaterializationBenchmark.graph_with_junctions_immediate           100  thrpt    5   58973,557 ±  2387,763  ops/s    98%
MaterializationBenchmark.sub_stream                                 1  thrpt    5  192419,870 ± 20621,692  ops/s    99%
MaterializationBenchmark.sub_stream                                10  thrpt    5  113372,461 ±  1416,131  ops/s    97%
MaterializationBenchmark.sub_stream                               100  thrpt    5   22558,577 ±   245,668  ops/s    90%


baseline
Benchmark                                                (complexity)   Mode  Cnt       Score       Error  Units
MaterializationBenchmark.flow_with_map                              1  thrpt    5  250048,147 ± 37096,379  ops/s
MaterializationBenchmark.flow_with_map                             10  thrpt    5  152853,149 ±  2257,300  ops/s
MaterializationBenchmark.flow_with_map                            100  thrpt    5   44561,527 ±   884,465  ops/s
MaterializationBenchmark.graph_with_imported_flow                   1  thrpt    5  237230,955 ± 18571,707  ops/s
MaterializationBenchmark.graph_with_imported_flow                  10  thrpt    5  134393,499 ±  7260,703  ops/s
MaterializationBenchmark.graph_with_imported_flow                 100  thrpt    5   33729,086 ±   500,537  ops/s
MaterializationBenchmark.graph_with_junctions_gradual               1  thrpt    5  231609,720 ± 19761,056  ops/s
MaterializationBenchmark.graph_with_junctions_gradual              10  thrpt    5  104245,973 ±   903,936  ops/s
MaterializationBenchmark.graph_with_junctions_gradual             100  thrpt    5    9551,597 ±   285,907  ops/s
MaterializationBenchmark.graph_with_junctions_immediate             1  thrpt    5  212552,446 ± 11135,513  ops/s
MaterializationBenchmark.graph_with_junctions_immediate            10  thrpt    5  161853,035 ± 11158,104  ops/s
MaterializationBenchmark.graph_with_junctions_immediate           100  thrpt    5   60006,484 ±   552,709  ops/s
MaterializationBenchmark.sub_stream                                 1  thrpt    5  195004,440 ± 17202,986  ops/s
MaterializationBenchmark.sub_stream                                10  thrpt    5  116940,618 ±  9665,168  ops/s
MaterializationBenchmark.sub_stream                               100  thrpt    5   24795,317 ±   307,664  ops/s
```