@johanandren 2021-01-11
@ygree this is a rehash of the other PR with stream operator source location, in this one we added a new attribute instead and updated a few `toStrings`. Cinnamon uses the `Name` attribute directly and not toString on GraphStage/GraphStageLogic, so this should be safe, right?