@johanandren 2021-01-11
I just realized the Materialization bench pre-constructs the blueprints and only materializes 🤦  so those numbers are not very relevant (the hit seems consistent though, guessing materialization is looking through the attributes somewhere, and the attribute list grew by one)