@johanandren 2021-01-11
I take no caching needed back, several orders of magnitude slower without the cache.

```
SourceLocation attribute
Benchmark                            (complexity)   Mode  Cnt        Score       Error  Units
GraphBuilderBenchmark.flow_with_map             1  thrpt    5  2006074,377 ± 46254,486  ops/s
GraphBuilderBenchmark.flow_with_map            10  thrpt    5   440028,273 ±  7369,840  ops/s
GraphBuilderBenchmark.flow_with_map           100  thrpt    5    49191,323 ±  1024,077  ops/s
GraphBuilderBenchmark.flow_with_map          1000  thrpt    5     4864,424 ±   540,625  ops/s

SourceLocation attribute, no cache
GraphBuilderBenchmark.flow_with_map             1  thrpt    5  915,667 ± 14,071  ops/s
GraphBuilderBenchmark.flow_with_map            10  thrpt    5   91,233 ±  1,945  ops/s
GraphBuilderBenchmark.flow_with_map           100  thrpt    5    9,122 ±  0,123  ops/s
GraphBuilderBenchmark.flow_with_map          1000  thrpt    5    0,908 ±  0,039  ops/s

baseline
GraphBuilderBenchmark.flow_with_map             1  thrpt    5  2168566,233 ± 69532,288  ops/s
GraphBuilderBenchmark.flow_with_map            10  thrpt    5   487062,918 ±  3871,123  ops/s
GraphBuilderBenchmark.flow_with_map           100  thrpt    5    57233,815 ±   863,505  ops/s
GraphBuilderBenchmark.flow_with_map          1000  thrpt    5     5764,395 ±    92,741  ops/s
```

I'll verify that the class makes sense, but it is problematic with a static cache for something that could potentially be generated dynamically (lambdas closing over some instance variable) and being stuck in the cache.