@johanandren 2021-01-11
I came up with something better. We only rarely actually use that string, so better create it when it is used and just keep a reference to the lambda in the attribute instead, since the attribute is attached to the stream operator it should be gc:d fine together with the operator when that is no longer used.

Much closer to baseline 👍 
```
[info] Benchmark                            (complexity)   Mode  Cnt        Score       Error  Units
[info] GraphBuilderBenchmark.flow_with_map             1  thrpt    5  2151397,423 ± 32125,165  ops/s
[info] GraphBuilderBenchmark.flow_with_map            10  thrpt    5   449591,545 ±  1568,294  ops/s
[info] GraphBuilderBenchmark.flow_with_map           100  thrpt    5    48872,650 ±  2825,224  ops/s
[info] GraphBuilderBenchmark.flow_with_map          1000  thrpt    5     5103,083 ±   675,149  ops/s
```