@jrudolph 2021-01-11
:+1: Thanks @johanandren for getting this idea into a mergeable state!